Purge rejected hetjob jobs at submission time
If a hetjob component failed job creation (e.g. limits) after successful
hetjob components the successful components would be sent to the database as
canceled and would later be purged according to MinJobAge.

This patch purges the successful pack components of a failed het job at
submission time. By doing this it avoids corrupted hetjobs where the
pack_job_id_set was NULL and the hetjob was never purged.

Bug 6847