Create common lua boilerplate code
Migrate a lot of common functionality out of job_submit/lua and
add to src/common/xlua.[ch].  Specifically the load_script,
register table, and job record table setup code is moved (more
of the job record table code could move as well).

This creates a mechanism for a plugin to call a centralized
xlua_load_script at every entrypoint to assure the latest lua
script is in use, while also providing a mechanism for the
plugin to perform further localization/customization to the
lua state before committing.  The behaviors of xlua_load_script
are well defined and avoids a few minor edge cases the previous
version in job_submit/lua exhibited.

The jobcomp/lua and job_submit/lua plugins are modified to use
this common code, and specifically can share the job_record
table code.  This will assure that both have the same programming
interface, and modifications to that interface benefit both.